requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.2
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
python-dotenv==1.0.0